Quite a lot of work here, probably to be honest it is still work in progress but I expect this represents and improvement on the old version. the object was to improve the quality of the autoz function, in particular allow improper dihedrals to be used as well as proper ones. This has been implemented as a function autoz within this class, incorporating code to define the internal tuples and to build up the z-matrix.
The find_i2 and find_i3  functions are now used to decide what
are valid z-matrix definitions (they are shared with code used
to implement the hybridise function and to convert individual
atoms to z-matrix form, so hopefully as these functions can
be smartened up together.

One aspect of teh change is that find_i2 and find_i3 both
now take an angle as an optional argument (testang= ) which
can be used to set how close an angle has to be to 90 before
it is allowable in a z-matrix definition.

also these functions can work when the atoms are out of
order provided the flag a.ok is set for each atom a that
has a valid definition (pass check=OK_CHECK) to request
this, otherwise atoms can only be defined from atoms above
them in the list.

There are also some enhancements to the import_geometry
function , correcting a problem with dihedral convention,
(still need to valididate this using gamess-uk chemshell
geometry optimisations). The code will now update variables
and constants if required from the incoming structure.