main change is a switch of old autoz function (implemented here) to new one (implemented in zmatrix.py).
also tried to fix a bug which meant that if you change i1,i2,i3
for an atom the required changes to r,theta,phi were not made
if r,theta,phi were variables. The reason behing this was that
variables maybe shared and so the update will cause problems,
so this case still has to be tested and should generate an error.
there are cases with a lot of unshared vars (such as generated
using the options within the editor) which need to work hence
this partial fix.