Have implemented an "All Visualiser" class for changing how all the Zmatrix-type objects are displayed at the same time. This inherits from the Visualiser class and has necessitated a few small changes to that class.
Have also implemented a "selector widget" to choose the frames for an animation and to manipulate their order. This resides in the file objects/selector.py. This works(-ish), it's still a bit flaky and has lots of scope for improvement, but needs to be tested to see where to best invest the effort in improving it.

The "Save Image" button in the File menu now brings up a dialog that allows the user to choose between hpeg and png image formats and specify the quality of the jpeg.

__balloon in main.py has been renamed to balloon so that it can be passed through for use by other widgets.

The View menu now greys out the animation , Show All etc options if there is nothing to display.

The animation controls have been revised to use the self.ani_list, which is the data structure that holds the list of images for the animation. The play button has been revised so that it plays from whereever the movie was last stopped.