Events are now not bound to tags, but the tag is worked out from the line number - this makes the clicking selections more robust.
Killing the widget with the window manger now does the same as the close button.

The graph.ani_list is now operated on directly so that the main widget and the selector are in sync. A number of callbacks have also been implemented to keep them in line.

A number of other small changes to the structure that will hopefully make it easier to move to the idea of multiple frames.

I think most of this works as I've tried clicking here there and everywhere without (many) disasters, however gui programming being what it is someone else should see if it behaves as expected...