Paul commiting Jens's smeagol reading changes (not finished but seem to be safe to commit)
lots of fixes to readers broken by change to quick_mol_view a while
ago. There should now be less updates to the screen when reading in
multiple molecules.


default for stick_type, sphere_type and line_type are now all 2, which
means using vtkPolyData with glyphs, tubefilters etc. This now allows
for colouring and sizing of spheres (and incidentally colouring of
sticks if chosen) following code in the newer VTK examples for
PDB/XYZ/Cube readers. This uses a vector array to hold the sizes (very
wasteful but seems the best option we have at the moment)
These viewers should be much cheaper for large molecules.


NB - now we get a traceback line for each Render(), this is to help track
unwanted refresh of screen.

Picture resets zoom to fit each time a molecule is read in (suggested
by Konrad Hinsen)