Fix to the subprocess.Spawn code. This wasn't working properly as stdin & stdout were being redirected with < and > symbols via the command supplied to the Spawn object. Unfortunately this doesn't work with the execvp call, as the redirect symbols get interpreted as arguments to the command being invoked.
The fix is to either pass in files and duplicate the file descriptors to the stdin,
stdout and stderr of the child process, or use a pipe to send the data back to the
parent.

This seems to work o.k., although there is a slight problem in getting stderr to
be displayed on job exit when the user supplies a file for stderr (the file handle
is either stale or of the incorrect type when the job finishes). However, as this
is only needed for debugging, I don't consider this a problem.