A few changes so that the gui will now save any changes made to the options (as defined in the Edit->Options menu - e.g. the background colour) to the ccp1guirc file, so that they are kept between restarts.
The format of the variables in the ccp1guirc file has changed slightly, so that instead of (e.g. self.bg_rgb = (0,0,100) ), the self. is not longer needed. So to set the path the dalton script, for example, the line would be:

dalton_script = 'home/jmht/dalton-2.0/bin/dalton'

If the variables in the ccp1guirc file are commented out, they will not be used and upadated values won't be saved out to the file.

If a ccp1guirc file doesn't exist, one will be created when the ccp1gui exits, containging all the exisiting values.

I have implemented an rc_vars dictionary that holds all of these variables that sits in this file and should be imported for use with:

from viewer.main import rc_vars