The FileTool wasn't working properly when given a command to invoke, as when it was called from the binding to the EntryField widget it was getting an event passed through, but when called from the FindFile (browser) method, no arguments were being passed through.
I've therefore tweaked it so that it passes the filename from the tkFileDialog as the argument when called from FindFile.

Unfortunately, this means that, if the command needs to query the value of the widget, it needs to check the argument to see what type it is (see __update_script in the dalton.py file for an example).