Have cleaned up the way that GAMESS-UK is invoked.
A get_runmethod function queries the rc_vars to see if it can find a rungamess script, failing this it searches for a rungamess in the system path/gui directories. If that fails it then qeuries the rc_vars for a gamess-uk binary and then searches the system path for one. This may be overkill, but it did actually find me an old binary I had totally forgotten about...

I've also cleaned up the __keepfiles method so that it either returns a command line suitable for rungamess, or sets environment variables to keep the relevant files.