Jens committing Paul's fix to the extend code.
self.cell is supposed to be a Scientific Vector and this is what was being
read in by the punch reader. However the code in extend assumed that self.cell
was a list and created vx by making a Vector from self.cell[0] etc. Although
this was incorrect it did not cause problems with earlier versions of Numeric & Scientific. A recent update to one of these modules means that this now causes an error, so the extra vector operation in the extend code has been removed.