Add the gui directory to the PYTHONPATH before we search for modules in case theuser places a module in the GUI directory.
Have also changed the code for reading in the rc_vars as this was messing up under windows when it tried to read in paths that had strings such as \b or \1 in them.

Because of this we need to change all strings to be raw strings so that the escape characters are not processed. The routine that converts a string to raw representation was nicked off the Python Cookbook website, as I couldn't work out a way to do this with the encode/decode string operators.