A number of small changes as the last output file I tried to read in crashed a number of the routines.
regexp in _read_orient_geom couldn't handle a symbol field with a number.

The list disp isn't initialised in the VibFreq object at init and so needs to be
declared for use in _read_normal_modes_force

The regexp in _read_frequencies_hessian wasn't robust enough and was missing the
fields it was supposed to be looking for.

_read_frequencies_force wasn't working with all outputs so I've used regular expressions
to make it a bit more robust.