Implemented a get_connected function that determines if an atom is part of a fragment (i.e. connected to the main molecule by only one bond) or whether it loops back on itself.
Also added a update_bond_distances function that updates the bond distance when the element type for an x-atom is set (obviously this only works if the x-atom is connected to at least one element of a defined type and not just other x-atoms).

Also changed the element symbols for the fragments so that they are upper case as this was confusing the symmetry detector.