Added the code to change bond lengths to something sensible when the user is building up a molecule and setting the element types of x-atoms.
Added the stuff to run the am1 optimiser. This is deactivated by default and the user should place the line "am1=1" in their .ccp1guirc.py file.

The ccp1guirc.py file is now called .ccp1guirc.py (i.e. prepended with a dot) so that it is hidden on unix systems.

Implemented a get_one_molecule function that returns either the first selected moleecule or the first moleucle in the data list if there is only one.

The control widget for the volume visualiser now appears by default when the user creates a volume visualistion object.