Added some bindings for mac osx darwin so that holding down the shift key and moving the mouse has the same effect as moving the mouse with the 3rd button held down under unix, and so that movig the mouse whilst holding the control key has the same effect as moving the mouse with the second button held down under unix.
Moving from vtk 4.? -> 5.0 also appears to change some things, so I've added a self.vtkVersion flag to check the vtk version and changed a font size and the type of an actor dependant on this change.

The volume visualiser was broken when a cube file was read in as there was no mapping present. I've fixed this, and also changed a test as to whether the axis was aligned to use the axis_aligned method of the Field class.