The rc_vars dictionary now lives in it's own file.
The way that the .ccp1guirc.py file is handled is also changed. On writing out
the variables, any new ones are appended to the point in the file where the
last variable was found.

On reading in the file, the gui now expects to see a line with:

###################### End User Defaults #####################

on it to signify the end of the defaults. Things will still work if this is
not present (so old files will still work), but strange variables could make
their way into the rc_vars if it is  not.

Also removed the get_one_molecule method, as there was already a choose_mol method
that I hadn't spotted. I've updated the choose_mol method to pick the currently
selected molecule if there is more than one molecule present, and only ask the user
if there is more than one molecule and none are selected.