A number of changes:
Specialised modules (e.g. SOAPpy, arclib) are only imported when they are needed.

The get_rcvars function is now available to all jobs and is used to update the job_parameters dictionary for the job.

'count' is now used instead of 'nproc' for the number of processors.

Added support for Nordugrid jobs. This is not finished yet, but does seem to work for the jobs that I've tried.