Have updated this to include a widget to work with Globus RSL strings.
The idea is that there is a list of supported variables and you can then add or remove them to a list of parameters for the current calculation and then edit them.

On closing the editor the user can either choose whether to save the parameters for just the current calculation, or save them as defaults, in which case they will be saved to the ccp1guirc file and will be used as defaults for all subsequent calculations.

This is currently only used for Nordugrid jobs.

This is a first shot at this. It worked when I tried it, but is likely to fall over if pushed. so please let me know where the bugs crop up.