The facility to edit Dalton input files was broken so this fixes it.
Have also enabled it so that an onsave command can be passed to the inputeditor. This should carry out any commands required to be carried out when the user hits save in the input editor and should return the name of the file to be written out.

In the case of dalton, this writes out the .mol file and returns the name of the .dal file. This is required so that when the user hits "Run InputFile" there is a valid .mol file to be used.