A few changes to the Nordugrid submission.
Python is not actually thread safe, and so there is a Global Interpreter Lock (GIL)
that ensures that only 1 thread can access python objects at a time. This lock needs
to be released and then re-acquired in any python code run in a thread or else the
whole interpreter will hang waiting for the operation to complete.

In the case of Nordugrid, this means that the GUI hangs when the brokering/submission
is being performed. This can be solved by the nordugrid developers placing the appropriate
calls to release the GIL in their code, or supply the relevant commands to SWIG to get it
to do this for them.

However, as this won't be activated in current releases, I've changed the Nordugrid stuff to
run in a separate process.

Unfortunately, this brings with it the problem that the fork call seems to monkey around with
the Globus authentication mechanism, so that jos are then rejected by the clusters. This
mainly seems to occur when an arclib "Certificate" call is made via Python in the main thread.

To work around this, I've therefore called "grid-proxy-info" to query the state of the user's
certificate in a separate process, thereby bypassing the python code.

This seems to work in most but not all cases.

In any case I've left things so that it will be trivial to revert to the original threaded code
as soon as the GIL issues are resolved.