Have changed how the base class subprocess handles commands. Instead of the cmd argument being a string that defines the command and all it's arguments, there is a cmd argument and an optional "args" keyword argument that provides the arguments as a list. This makes it possible to deal with arguments to a command that have embedded spaces, which otherwise a decent into quoting hell...
This change has required quite a few changes through the rest of the code.

Have added a cmd_as_string method that returns the full command as a string for those methods that still need it.

Fixed a bug in the wait method that only kept the output if debugging was turned on.

Added a FAILED jobstatus so that we can tell when a job has died.

Added get_output and get_error methods to the Spawn class as this what is used to invoke the grid tools for Growl.