Firstly a few changes relating to the change to the SubProcess class, which now requires that all commands are specified as a command string and a list of arguments. This also entails adding local_command_args and remote_command_args to the JobStep class.
Have created a GridJob class that should hold all methods that are common to all GridJobs. As I'm not entirely sure what this should be yet it's relatively empty... The CreateRSLString method is used to build up an RSL string from the job_parameters. The CheckProxy method uses grid-proxy-info to check if the grid-certificate is valid.

Have added a GrowlJob class for running jobs on the grid using Growl. This has been tested for serial and parallel jobs on scarf.rl.ac.uk and not alot else. It'll only take a small shove to get this to fall over, but I need to know what are the most likely shoves are so that I can start hardening it up, so please give it a try!