first stage of a tidy up of the process/j handling. quite a lot of redundant code has been moved out of subprocess.py into slaveprocess.py unit tests are being added for the different sub-process handling classes and will be added for the job manager stuff in due course. This should simplify the maintenance of the code and provide an up-to-date (ie regularly checked) usage example also
this change introduces a LocalJob class (to replace ForegroundJob
and BackgroundJob) and also introduces RemoteJob and RemoteBatchJob
(currently being implemented) to provide better support for interfacing
with queueing systems.

************************** health warning *****************************
It is recommended to delay cvs updates for a couple of days unless
you are keen to help sort this out. Repository should be stable again
by the end of the week.