Added definiting of LOCALHOST so that the hostname of the machine can be used when displaying the submit menu.
Added several variables to the base job class

poll_interval - as we always need to wait for a couple of intervals when the user presses suspend to see if it is possible to stop the job.
stopjob - flag set by the job manager to indicate the job should be stopped
restart - set if a job has been halted mid-step and so (e.g.) should go straight to the polling loop rather then submitting a job when it is restarted
thread - used to hold the thread the job is run under so that the jobeditor can control it

Also changed things so that each job should supply a _kill method as the main kill method sits in the base job class

Have enabled restarts for Nordugrid and Growl jobs sofar.

A few other minor fixes.