symmetry off by default (just too many problems with point grp changes).
support for adding keywords onto the end of z-matrix variables
e.g. "type 3"

support for performing a hessian calculation before running a transition
state

support for potential derived charges (RHF only at the moment)

recover the structure from a
z-matrix based optimisation in such a way that the z-matrix is
retained and can be edited and re-submitted.

output of mulliken/lowdin charges to punchfiles