The setup.py script is used by disutils and can be used to install the ccp1gui into the python distribution on the machine. This isn't particularly useful by itself, but was required by py2exe, which I was using to create a stand-alone windows distribution that could be used on machines that didn't have python installed on them. This was required as a user wanted to run the CCP1GUI/GAMESS-UK off a CD so that they could hand this out to students in a practical and allow then to run on the machines in the teaching lab without installing any software on them.
I'm committing these files as this approach might also work on Linux and so could become quite handy in the future.

Unfortunately, the setup.py script doesn't work without hacking the distutils code on the host machine, as it appears to have a bug in the way directories are referenced (see notes in the setup.py file).

The __init__.py script is required to get the ccp1gui directory treated as a package by the script.