Changes to enable the gui to read in a sequence of structures in an .mol xyz file as a trajectory.
Currently this is only enabled for files with the suffix ANI

This has required updating the rdxyz method in main.py so that it takes an optional argument "sequence", which determines whether the structures are being read in as individual molecules or as a trajectory.

Have also added a add_molecule method to the ZMatrixSequence class which updates the atom list for the ZMatrixSequence with the atom list of the first molecule added, as I got rather confused trying to work out why it wasn't working for me when I first tried.