A number of changes to make the handling of writing input files and runnig calculations a bit more sensible.
Now there is no "Run Inputfile" method, but the makejob method should make a check on whether the input file exists and then decide on whether to use an existing one (which may have been edited) or create a fresh one.

The WriteInput method of the Calced object now informs the user of the path to the input file it has written as requested by Joop.

I've added a createInput method for GAMESS-UK that the InputEditor attempts to invoke to create an input as a list in case an input file hasn't already created.

This has necessitated changing the way that GAMESS-UK creates it's input, so that it is now created as a list and only written out by the WriteInput method.

The Dalton interfaces has been ammended to remove some dead code (left there from when I copied stuff over from the GUK interface) and so that it returns the names of the inputfiles it has written.

Molpro interface also ammended and a bug with the basis set naming fixed.