A script to take an old GAMESS-UK output file and create a new input file with the same directives as the original input file, but with the geometry specification taken from the last geometry given in the ouptut.
The name of the input file is derived from the output, replacing .out with .in.

The script could easily be made to work from a punch file and the corresponding input file, but this hasn't been enabled yet.

The script takes a -z or --zmatrix flag which indicates that it should output the geometry as a zmatrix, otherwise it defaults to cartesian. Currently if a zmatrix is requested, but a zmatrix could not be found in the output, this is considered an error and the script will bomb.