Forgot to  describe the changes to this file...
Have moved the RSL stuff into a separate class that the grid jobs can inherit from as
the other classes that don't require this were being made needlessly complex by it's inclusion.

Have also cleaned up the way that the initial parameters are read in and written out to the rc_vars file so that variables (e.g. working directories/executable locations) are kept in separate namespace but operated on by the same tools.  For this reason, things like the rc_vars dalton_script is now named DALTON___executable and the GAMESS-UK equivalent is GAMESSUK___executable.

This change allows the jobed to be used to configure the working directories and script locations, removing this code from the calculation-specific classes and makeing the interface more constent across the different codes.

The LayoutDirectoryWidget and LayoutExecutableWidget methods now have an optionl "browse" parameter that indicates if their location is expected on the local filesystem, in which case there is now a "Browse"button to bring up a file browser to find the location.

The relevant widgets also take an optional argument chdir_cmd or chexe_cmd that can be passed in and is called each time the widget is used, so that the directory/executable parameter of the calculation is updated in realtime.