Changed things so that the we now use the job editor to configure the location of the executable and working directory.
There are now no separate rc_variables for rungamess and the gamess binary, there is just the one and we work out which we are using from the name of the executable (i.e. if it contains rungamess, we assume it is the rungamess script, otherwise we assume it is the binary.

Also changed the warning message for when the user tries to run a job where an input file already exists so that the default action is to overwrite the file.