Have just discovered that when a GUK opti run crashes with a bad zmatrix error the final geometry printed in the output is the one that caused the crash and hence cannot be used to restart an optimisation.
The script has therefore been extended to use the punch file to get the geometries.

This necessiates supplying a -i or --input-file argument that causes the script to
use the file that follows as the template input file (i.e. the that will have the
new geometry inserted into it).

The -i flag can also be used if a GUK output file is used so that the script can be
used to put a new geometry into an arbitrary input file.