The code for delete_unselected was broken as it only deleted the atoms, not the connectivity table. When the GUI tried to build the molecule it was therefore expecting to draw a bond to an atom that no longer existed and throwing an exception.
I've changed the code for delete_unselected so that it now just calls delete_atom (which is
called by "Delete Selected Atoms") but with an additional "unselected" flag. This causes the
selection of the atoms to be inverted in this routine. Having both calls use the same code
will hopefully stop the methods diverging again and also means that delete_unselected now has
an undo function.

This checkin also changes the GUI's currently default behaviour to remember the last directory
it was used in to be dependant on the user setting the "old_path" in their .ccp1guirc.py file.