This is the first of a bunch of checkins that constitute a big change to the way that we deal with reading and writing files and enables us to use OpenBabel to read and write files if the python interface is available in the path.
There are now separate objects that handle both reading and writing for particular
filetypes. These objects need to supply a _ReadFile method for reading a file that
builds up a list of object types they have read (e.g. molecules, fields etc) and
a (e.g.) _WriteMolecule method to write out a molecule to a particular file.

They should all inherit from the FileIO object (file: interfaces/fileio.py) that provides
a number of methods that should be useful to all IO objects (e.g. splitting up the filename,
returning objects etc).

The file interfaces/getfileio.py contains a GetFileIO object that can be called to determine
the correct reader for a particular file. This object contains a data structure called format_info
that contains information on the capabilities of the particular IO objects. This isn't dynmically
populated so anyone who adds new readers/writers or updates their capabilities needs to ensure that
this structure is update.

There is more documentation in the files - if any of it isn't clear of anyone has trouble using it
please let me know.