Reinstated the 'Molecules' type at the list of available filetypes in the "Open" dialog.
Also a number of small changes that enable the testing of the openbabel IO stuff.

There is currently an issue in that if openbabel is available and works some of the other
tests will fail as they just check for the size of a file that is written and for (e.g. pdb)
this is different depending on which IO object writes the file. To be fixed soon...