Changes to address the mounting confusion that was arising from the interaction of jobs, calculation, calcualtion editors and job editors.
The calculation now holds the job for most of these operations and there are create_job, set_job and get_job methods for accessing it.

The calculation editor has similar methods but these just wrap the calc class methods.

The calculation editor now holds a dictionary that remembers the parameters for all jobs that have been edited during a particular session ( this is separate from any parameters that are saved in the defaults).