Enable the gui to plot 'streamarrows'. These are arrows that are placed at points calculated along the streamline path. Combined with the streamline representation, this allows the gui to plot flow lines that can be used to visualise (e.g.) magnetic field vectors.
As part of this work, I've enabled a 'thin points' option for both streamlines and
streamarrows. This takes the sample grid and sub-samples a subset of the points so
that far fewer paths are calcualted. I've found this handy for testing out
different time steps/integration times etc quickly before reverting to the full grid
and have computer grind away for ages calculating all the paths. This options is also
useful for quickly reducing the number of arrows on the plots, as they can easily
get out of control.

I've also had to change the widgets so that the tools to configure the streamlines,
streamarrows and the grid editor can be hidden unless the user clicks the relevant 'display'
option - this is because the widget with all tools displayed is too large to display on
the screen on my Mac and there is then no way to access the buttons at the bottom.

There is some more work to be done on this - something that I'm going to try and get round
to is to put some more intelligence into deciding what needs to be redrawn/recomputed,
as currently the entire visualisation pipeline is redone in full each time the user
hits 'Update' which for big visualisations really slows things down.