Quite a few things...
Made the code to draw the streamarrows a bit more polished so that it's now possible to select between
arrows and cones with the widget.

Changed the way that the visualisers handle colour. There is now a Colourer object that holds all the information about how things are coloured. Created a ColourChooser widget that acts on this and sets its attributes. This allows a single widget to be used to set how things are coloured - which is handy for the vector visualisers where you can colour by a single colour, an attribute of the vector or by a separate scalar field.

Have also enabled "colourmap_actors" - these are bars that appear on the sides of the screen and show the colourmap and range that is being used to colour the object. I've found this really handy in getting good pictures as it's really easy to see how to tweak the parameters to get things looking good. Turning this on is done by selecting the "display" variable in the ColourChooser widget.

As part of this I've done a bit of a hatchet job on the visualiser code, and the vector visualiser code in particular, but I hope it's an improvement.

This code has already been alpha tested by two users who were interested in the vector plotting stuff so I'm hoping that I'm not going to break anything too major with this...