Files for embedding the ipython shell in the gui (experimental).
I've taken some of this stuff from the accercise project who have written stuff to embed ipython in a GTK2 program and have rewritten it to use Tkinter/Pmw. However a retrograde step that might seem... ;-)

This stuff is under a BSD licence, which is included in the COPYING file.