Enabled the ability to save images as TIFFs and have also put the slider to set the resolutuon of jpegs as Fedor has said he would like this.
The code was (is..) a bit of a mess so I've cleaned it up and reinstated the old code that used the Pmw dialog.

The code added by Paul that used the tk file dialog is still there, but deactivated by default. To turn it on again set save_image_dialog_quick to True in the .ccp1guirc.py file.

I've only tested this on Mac so it might need a bit of tweaking to get it working on Linux/Windows.