This is one of python files from the VTK project, which we've had to add due to a problem with how tk objects are handled in Python 2.3:
http://www.vtk.org/pipermail/vtkusers/2008-September/096743.html

We can't use the Tkinter.wantobjects=0 fix as it breaks other stuff within Tkinter.

Although this supposedly appeared in Py 2.3, it doesn't seem to have made it into any packaged version before 2.6

The problem was fixed in the vtkLoadPythonTkWidgets file on 03/07/2009 which corresponds to vtk 5.4

There is a conditional import for this file in the file:

viewer/vtkTkRenderWidgetP.py

so we only import this version if it is needed.