First of a number of commits to tidy up the gui and implement some testing in preparation to removing the dependencies on Numeric an d Scientific Python.
Changed the import of Scientific python so that it just imports the module and doesn't do an "from x import *". This will make it easiser and clearer to spot where it's being used. In future I'll try and remove all "from x import *" statements as it's bad practice and confusing.

Removed the "apply" statements as they are deprecated.

Removed numerous print statements or turned them into debug statements so that they don't clutter the output where running under tes
ting.

Changed the density visualisers to be based on the DataVisualiser class so that the data summary widget is displayed.

Changed how the streamline display type is propogaed to make it easier to run the testing framework and test the different visualise
rs.