Changed the import of maths to make it explicit which attributes are being used and commented out the import of string as the module is deprecated.
Changed how the visualise function works and how the menus for the visualisers work. The hideous nested lambda function is now history and the visualise function can take either visualiser instance or the name of a visualiser. As well as being cleaner, this allows us to test the visualisers from within the unittest module.

Added a "display" flag to the load_from_file and import_view_objects so that it's possible to choose whether to display objects when they are imported.

Removed a bunch of print statements.