File to test the Molecule, Field and Vector visualisers.
Currently pretty basic and a bit hacky to work around issues to do with Tkinter and how the unittest framework expects to run things
.

Each test is expected to be entirely self-contained and a new instance of the test class is instantiated for each individual test. As a result, the setUp and tearDown methods are executed for each test. This means that, even if it would work, a fresh gui instance
would be created for each test, which is excessive, as all of the imports and readings of files would be repeated for testing the di
fferent visualisers on identical data.

However, even if we wanted to do this I can't work out how it would be done, because, as the test is run as a single python script,
we only have one root Tk instance, which would be destroyed when the first test was torn down. Subsequent tests therefore don't have
 a valid Tkinter instance to instantiate the gui with.

As a result, I've hacked things so that we create a global Tkinter.Tk() root instance which is shared by all the tests, and the gui
instance is shared as a class variable between the different tests within a class.