These are files that contain wrappers for (currently) the Numeric and Numeric LinearAlgebra libraries.
All calls to numerical functions should go through these files. The files currently try to import the libraries and if they are not present, exceptions are raised if the code tries to call these functions.

They also have an isAvailable attribute that can be used to check if we need to disable certain functionality if the libraries are not present.

Currently they only call Numeric, but I'm going to add wrappers for numpy so that we will be agnostic to whether Numeric or numpy is present.

As Numeric is (very) deprecated, the wrappers will try to be as close to the numpy functionality as possible, because presumably, if we start to use them more, we are more likely ot be reliying on numpy then Numeric.