Previously we used a modified version of the vtkTkRenderWidget file, which had been copied into our repository and then modified to create the Tk widget that we rendered into.
This has the disadvantage that it's not obvious what differs in our implementation from the default, and also that we do not take advantage of any updates/fixes to the file released by vtk.

This file therefore subclasses the vtkTkRenderWidget class with our own vtkTkRenderWidgetCCP1GUI class, which is modified appropriately.