A few fixes so that a single FileIO object can be used consequtively with multiple files.
To do this we need to nuke the objects from the previous read when we read in a new file.

It is now possible to just call GetObjects with a different filepath parameter to read in a different file and return the objects.

Also removed the GetClass method as we can use that which is now implemented in the objects themselves.