Several changes.
We now use the GetClass method of objects, which cleans up the code a bit.

I've removed the quick_mol_view and quick_trajectory_view methods and replaced them with a quick_obj_view, which takes a list of objects and a list of the types of objects that should be viewed. I've also added an "open" keyword which determines whether the visualiser widget should be opened-this is set to 1, so a visualiser is opened for each object added.

Also implemented a get_visualiser method which returns a visualiser for an object, as this functionality was required in a couple of places.