Large change to the layout of the code as part of changing how we handle picking when we are using sphere_type=2
All of the rendering code for drawing a different representation is now in a separate method, e.g. drawing sticks with type=2 is now in the method _build_sticks_typ
e_2. This makes the code more self contained and far easier to understand.

The wire/stick/sphere type 2 code all works off the same vtkPolyData, so this is created in the method _get_molecule_polydata, which should only be called once duri
ng each rendering pass, saving the data in the variable molecule_polydata so that it can be used by the different methods.

The picking code in mypick2 has been altered for sphere_type 2 so that it determines which atom has been picked from the InputPointIds of the
vtkGlyph3D object.

Changing the code now means that picking is enabled for both wire and sphere representations (it was broken for spheres) and we can now
display sticks of type 2 without requiring the wire representation as well.

There is almost certainly more work to do on this and I'm not convinced all the bugs/performance issues have been shaken out, but
picking is at least working on my machine now which it wasn't previously.