Fix for bug 2992807 on sourceforge:
"New Molecule => Add tBu or Et group: The bonds in the new group are messed up, it looks like almost all atoms are interconnected"

The problem was that in the add_fragment method, connect  and calculate_coordinates were called on the zfragment. This couldn't work as the zmatrix definition was incomplete, so calculate coordinates returned several coordiates as 0 and the connect routine ended up connecting everything to everything else.

Moving the connect call to after the fragments have been joined seems to fix the problem.