A couple of small fixes to the test cases.
We need to leave a 2-second gap between running the different cases as otherwise the gamess or dalton process doesn't have a chance to clear up after itself before the next one is invoked.

Also changed things so that the dalton calculation looks for a dalton script in the path.